refactor: Retrofit Client 제거
